+++ build.rs
@@ -1,8 +1,7 @@
 use std::{env, process::Command};
 
-const CLEARANCE_NAME: &str = "clearance";
-const CLEARANCE_PKG_DIR: &str = "ofdb-app-clearance";
-const CLEARANCE_PKG_SRC: &str = "ofdb-app-clearance/src";
+const CLEARANCE_APP_DIR: &str = "ofdb-app-clearance";
+const CLEARANCE_APP_SRC: &str = "ofdb-app-clearance/src";
 const CLEARANCE_FEATURE_NAME: &str = "clearance";
 
 fn main() {
@@ -12,21 +11,14 @@ fn main() {
     ))
     .is_ok()
     {
-        assert_wasm_pack_is_installed();
-        Command::new("wasm-pack")
-            .args(&[
-                "build",
-                "--target",
-                "web",
-                "--release",
-                "--out-name",
-                CLEARANCE_NAME,
-                CLEARANCE_PKG_DIR,
-            ])
+        assert_trunk_is_installed();
+        Command::new("trunk")
+            .args(&["build", "--release"])
+            .current_dir(CLEARANCE_APP_DIR)
             .status()
-            .expect("Unable to successfully execute wasm-pack");
-        println!("cargo:rerun-if-changed=\"{}\"", CLEARANCE_PKG_SRC);
-        for entry in walkdir::WalkDir::new(CLEARANCE_PKG_SRC)
+            .expect("Unable to successfully execute trunk");
+        println!("cargo:rerun-if-changed=\"{}\"", CLEARANCE_APP_SRC);
+        for entry in walkdir::WalkDir::new(CLEARANCE_APP_SRC)
             .into_iter()
             .filter_map(|e| e.ok())
         {
@@ -35,16 +27,16 @@ fn main() {
     }
 }
 
-fn assert_wasm_pack_is_installed() {
+fn assert_trunk_is_installed() {
     let output = Command::new("cargo")
         .args(&["install", "--list"])
         .output()
-        .expect("Unable to check wasm-pack installation");
+        .expect("Unable to check trunk installation");
     let output_string = String::from_utf8(output.stdout).unwrap();
-    if !output_string.contains("wasm-pack") {
+    if !output_string.contains("trunk") {
         Command::new("cargo")
-            .args(&["install", "wasm-pack"])
+            .args(&["install", "trunk"])
             .status()
-            .expect("Unable install wasm-pack");
+            .expect("Unable install trunk");
     }
 }