+++ ofdb-app-clearance/src/main.rs
@@ -57,7 +57,6 @@ fn view(mdl: &Mdl) -> Node<Msg> {
     }
 }
 
-#[wasm_bindgen(start)]
-pub fn start() {
+pub fn main() {
     App::start("app", init, update, view);
 }