+++ shell.nix
@@ -2,7 +2,7 @@ let
   moz_overlay = import (builtins.fetchTarball https://github.com/mozilla/nixpkgs-mozilla/archive/master.tar.gz);
   pkgs = import <nixpkgs> { overlays = [ moz_overlay ]; };
   rustChannel = pkgs.rustChannelOf {
-     date = "2021-05-05";
+     date = "2021-08-06";
      channel = "nightly";
   };
   rust = (rustChannel.rust.override {